// src/app/api/fn/inventory/inventory-adjustments-list.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { PaginatedAdjustmentList } from '../../models/paginated-adjustment-list';

export interface InventoryAdjustmentsList$Params {

/**
 * On or after this date.
 */
  date_from?: string;

/**
 * On or before this date.
 */
  date_to?: string;

/**
 * A page number within the paginated result set.
 */
  page?: number;

/**
 * Adjustment reason.
 */
  reason?: 'DAMAGE' | 'LOSS' | 'OPENING_BALANCE' | 'RETURN_TO_SUPPLIER' | 'SHOP_USE' | 'SUPPLIER_REPLACEMENT' | 'WARRANTY_REPLACEMENT';

/**
 * Matches codes, names and numbers.
 */
  search?: string;

/**
 * Document status.
 */
  status?: 'DRAFT' | 'POSTED';

/**
 * Supplier id.
 */
  supplier?: number;
}

export function inventoryAdjustmentsList(http: HttpClient, rootUrl: string, params?: InventoryAdjustmentsList$Params, context?: HttpContext): Observable<StrictHttpResponse<PaginatedAdjustmentList>> {
  const rb = new RequestBuilder(rootUrl, inventoryAdjustmentsList.PATH, 'get');
  if (params) {
    rb.query('date_from', params.date_from, {});
    rb.query('date_to', params.date_to, {});
    rb.query('page', params.page, {});
    rb.query('reason', params.reason, {});
    rb.query('search', params.search, {});
    rb.query('status', params.status, {});
    rb.query('supplier', params.supplier, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<PaginatedAdjustmentList>;
    })
  );
}

inventoryAdjustmentsList.PATH = '/api/inventory/adjustments/';